Create Subarray with given sum
The given code finds a sub-array in an unsorted array of non negative integers that has sum equal to s. It returns the left and right indexes (1-based) of that sub-array, or -1 if no such sub-array exists. If there are multiple subarrays, it returns the first one encountered while moving from left to right.